refactor: move campaigns list into external-rewards repo
