+++ packages/external-rewards/src/index.ts
@@ -1,3 +1,13 @@
-export { default as campaignList } from './campaign-list.json'
-export { default as rewardTags } from './reward-tags.json'
-export * as campaignsJsons from './campaigns'
+import { default as campaignList } from './campaign-list.json'
+import * as campaignsJsons from './campaigns'
+
+const parsedCampaignsJsons: { [key: string]: any } = campaignsJsons
+const campaigns = campaignList
+  .map(({ campaign }) => {
+    const campaignName = campaign.split('.')?.[0]
+    if (!campaignName || !(campaignName in parsedCampaignsJsons)) return null
+    return parsedCampaignsJsons[campaignName]
+  })
+  .filter((campaign) => campaign !== null)
+
+export default campaigns
